+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.0.0
+* `Text.Show.Text` now exports `Data.Text.Lazy.Builder` for convenience
+* Add `showLazy`, the counterpart of `show` for lazy `Text`
+
 # 0.2.0.0
 
 * Add `Show` instances for strict and lazy `Text`